+++ modules/enclave-commands/src/light_client.rs
@@ -22,7 +22,7 @@ pub struct InitClientInput {
     pub any_client_state: Any,
     #[serde(with = "AnyDef")]
     pub any_consensus_state: Any,
-    pub current_timestamp: u64,
+    pub current_timestamp: u128,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -30,7 +30,7 @@ pub struct UpdateClientInput {
     pub client_id: ClientId,
     #[serde(with = "AnyDef")]
     pub any_header: Any,
-    pub current_timestamp: u64,
+    pub current_timestamp: u128,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
